requirements.txt
flask
python-escpos
evdev
pyusb
flask-sqlalchemy
sqlalchemy
pycups
